+++ src/kernel/vfs.h
@@ -24,7 +24,7 @@ typedef struct VFSStat {
 typedef struct VFSInterface {
   void* (*open)(void* fs_data, const char* path, unsigned mode);
   size_t (*read)(void* fs_data, void* file, void* buffer, size_t size);
-  int (*write)(void* fs_data, void* file, const void* buffer, size_t size);
+  size_t (*write)(void* fs_data, void* file, const void* buffer, size_t size);
   int (*close)(void* fs_data, void* file);
   int (*seek)(void* fs_data, void* file, size_t offset);
   int (*mkdir)(void* fs_data, const char* path);
@@ -50,7 +50,7 @@ typedef struct VFSFileDescriptor {
 int vfs_mount(const char* path, VFSInterface* fs, void* fs_data);
 VFSFileDescriptor* vfs_open(const char* path, unsigned mode);
 size_t vfs_read(VFSFileDescriptor* fd, void* buffer, size_t size);
-int vfs_write(VFSFileDescriptor* fd, const void* buffer, size_t size);
+size_t vfs_write(VFSFileDescriptor* fd, const void* buffer, size_t size);
 int vfs_close(VFSFileDescriptor* fd);
 int vfs_seek(VFSFileDescriptor* fd, size_t offset);
 int vfs_mkdir(const char* path);
